réorganisation des fichiers en quelque chose de plus modulaire et ajout d'un canvas que j'ai fait dans un autre projet
