+++ assets/css/index.css
@@ -1,4 +1,5 @@
 body {
+    margin: 0;
     font-family: "Quicksand", sans-serif;
     max-width: 100%;
     max-height: 100vh;
@@ -19,18 +20,6 @@ canvas {
     height: 100%;
     pointer-events: none;
     z-index: -1;
-}
-
-a {
-    position: absolute;
-    bottom: 2vmin;
-    right: 2vmin;
-    color: rgba(255, 255, 255, 0.2);
-    text-decoration: none;
-}
-
-a:hover {
-    color: #fff;
 }
 
 /* Alert */
@@ -149,6 +138,8 @@ a:hover {
     padding-left: 5%;
     color: rgba(255, 255, 255, 0.737);
 }
+
+/* Delete task */
 
 #deleteArea {
     width: 50px;
