+++ assets/js/modules/clearTask.js
@@ -1,0 +1,7 @@
+// fonction pour supprimer tout les données du localstorage (bouton reset)
+
+export function clearTasks() {
+
+    localStorage.clear();
+  
+  };
