+++ index.html
@@ -14,7 +14,8 @@
 </head>
 
 <body>
-    <canvas></canvas>
+    <canvas id="canvasDark"></canvas>
+    <canvas id="canvasLight" hidden></canvas>
     <div id="alertEnterTask" class="alert alert-primary position-absolute" role="alert" hidden>
 
     </div>
@@ -80,7 +81,8 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-    <script src="./assets/js/script.js"></script>
+    <script type="module" src="./assets/js/script.js"></script>
+    <script type="module" src="Canvas/assets/js/script.js"></script>
 </body>
 
 </html>
